requirements.txt
django
djangorestframework
django-colorfield